feat: 序列化器与 SPI 机制
新增序列化器与 SPI 机制（Hessian存在问题）

+++ example-consumer/src/main/java/com/sora/example/consumer/EasyConsumerExample.java
@@ -22,7 +22,7 @@ public class EasyConsumerExample {
         } else {
             System.out.println("user == null");
         }
-        long number = userService.getNumber();
+        short number = userService.getNumber();
         System.out.println(number);
     }
 }
